requirements.txt
pypinfo
PyGithub
